+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.40 (2022-12-19)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.39](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.38...@parca/grafana-panel@0.0.39) (2022-12-19)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## [0.0.38](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.37...@parca/grafana-panel@0.0.38) (2022-12-14)
 
 **Note:** Version bump only for package @parca/grafana-panel